requirements.txt
opencv-python
numpy
pillow
pdf2image
tqdm